Add create poeme

+++ src/Zhongfu/BlogBundle/Controller/AdminArticleController.php
@@ -109,27 +109,29 @@ class AdminArticleController extends Controller
         /**
          * @var $oArticle Article
          */
-        foreach ($oArticle->getImagesList() as $oImage) {
+        if ($oArticle->getAvatar()) {
+            foreach ($oArticle->getImagesList() as $oImage) {
 
-            if (file_exists($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/' . $oImage->getFile())) {
-                unlink($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/' . $oImage->getFile());
+                if (file_exists($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/' . $oImage->getFile())) {
+                    unlink($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/' . $oImage->getFile());
+                }
+                if (file_exists($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile())) {
+                    unlink($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile());
+                }
+                $oEntityManager->remove($oImage);
+            }
+            //and the avatar
+            if (file_exists($this->container->getParameter('upload_base_dir') . $oArticle->getAvatar()->getPath() . '/' . $oImage->getFile())) {
+                unlink($this->container->getParameter('upload_base_dir') . $oArticle->getAvatar()->getPath() . '/' . $oImage->getFile());
             }
             if (file_exists($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile())) {
                 unlink($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile());
             }
+
+            rmdir($this->container->getParameter('upload_base_dir') . $oImage->getPath());
             $oEntityManager->remove($oImage);
         }
-        //and the avatar
-        if (file_exists($this->container->getParameter('upload_base_dir') . $oArticle->getAvatar()->getPath() . '/' . $oImage->getFile())) {
-            unlink($this->container->getParameter('upload_base_dir') . $oArticle->getAvatar()->getPath() . '/' . $oImage->getFile());
-        }
-        if (file_exists($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile())) {
-            unlink($this->container->getParameter('upload_base_dir') . $oImage->getPath() . '/mini_' . $oImage->getFile());
-        }
 
-        rmdir($this->container->getParameter('upload_base_dir') . $oImage->getPath());
-
-        $oEntityManager->remove($oImage);
         $oEntityManager->remove($oArticle);
         $oEntityManager->flush();
